requirement.txt
langchain==0.0.208
deeplake==3.9.27
langchain-google-genai
langchain
tiktoken==0.4.0
selenium==4.15.2
requests
pypdf
requests
newspaper3k